# Conflicts:
#	ui/src/main/java/com/geekbrains/team/filmlibrary/fragments/mainScreen/MainScreenViewModel.kt
#	ui/src/main/java/com/geekbrains/team/filmlibrary/fragments/mainScreen/model/UpcomingMovieView.kt
#	ui/src/main/java/com/geekbrains/team/filmlibrary/fragments/top/TopFragment.kt
#	ui/src/main/java/com/geekbrains/team/filmlibrary/fragments/top/TopMovieViewModel.kt
#	ui/src/main/java/com/geekbrains/team/filmlibrary/model/MovieView.kt
#	ui/src/main/java/com/geekbrains/team/filmlibrary/model/TVShowView.kt

+++ domain/src/main/java/com/geekbrains/team/domain/tv/model/TVShow.kt
@@ -26,7 +26,7 @@ data class TVShow(
     val originalLanguage: String,
     val originalName: String,
     val overview: String,
-    val popularity: Double,
+    val popularity: Int,
     val posterPath: String,
     val productionCompanies: List<ProductionCompany>? = null,
     val seasons: List<Season>? = null,
